// src/typings/enums.ts
export enum Routes {
	Login = '/',
	Register = '/sign-up',
	Settings = '/settings',
	SettingsEdit = '/settings-edit',
	Messenger = '/messenger',
};

export enum User {
	id = 'id',
	first_name = "first_name",
	display_name = "display_name",
	second_name = "second_name",
	login = "login",
	email = "email",
	phone = "phone",
	avatar = "avatar",
}